scholarmind/requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.5.0
tqdm==4.66.4

# LLM & Agents
langchain==0.2.11
langchain-community==0.2.10
langchain-anthropic==0.1.20
langchain-openai==0.1.17
langgraph==0.1.14
anthropic==0.32.0
openai==1.40.0

# RAG & Embeddings
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
torch==2.3.1
transformers==4.43.0

# Knowledge Graph
neo4j==5.22.0

# Data sources
arxiv==2.1.3
requests==2.32.3
httpx==0.27.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1

# API & UI
fastapi==0.111.1
uvicorn[standard]==0.30.3
streamlit==1.36.0
plotly==5.22.0
networkx==3.3

# Evaluation
ragas==0.1.10
datasets==2.20.0
pytest==8.2.2
pytest-asyncio==0.23.7

# Dev
black==24.4.2
ruff==0.5.2
ipykernel==6.29.5
